refact: Introduce Write Quill분리 및 코드 정렬
