+++ src/components/Introduce/Quill/index.tsx
@@ -3,36 +3,35 @@ import ReactQuill from 'react-quill';
 import 'react-quill/dist/quill.snow.css';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { AxiosError } from 'axios';
 import { QueryKeys, restFetcher } from '@/queryClient';
 import getImageUrls from '@/utils/Quill/getImageUrls';
 import { PostBoardAPI } from '@/apis/boards';
+import { uploadFile } from '@/apis/uploadS3';
 import useQuillModules from '@/hooks/useQuillModules';
 import { BoardDetailData } from '@/types/boardDetailType';
 import { BoardForm } from '@/types/boardType';
-import { freeCategory, advertiseCategory } from '@/constants/category';
 import styles from './styles.module.scss';
-import Title from '../Title';
-
-type CommunityQuillProps = {
-  queryParam: string | undefined;
-};
 
 // TODO: 이미지 10개 이상 등록 불가
 
-export default function CommunityQuill({ queryParam }: CommunityQuillProps) {
+export default function IntroduceQuill() {
   const navigate = useNavigate();
   const location = useLocation();
   const boardData: BoardDetailData | null = location.state;
 
   const QuillRef = useRef<ReactQuill>();
+  const thumbnailRef = useRef<HTMLInputElement>(null);
 
   const [title, setTitle] = useState(boardData ? boardData.title : '');
   const [contents, setContents] = useState('');
   const [category, setCategory] = useState(boardData ? boardData.category : '');
-  const prefixCategory =
-    queryParam === 'free_board' ? 'DEFAULT' : 'ADVERTISEMENT';
-  const categoryList =
-    queryParam === 'free_board' ? freeCategory : advertiseCategory;
+  const [thumbnail, setThumbnail] = useState(
+    boardData ? boardData.imageUrls[0] : '',
+  );
+  const [thumbnailTitle, setThumbnailTitle] = useState(
+    boardData ? boardData.imageUrls[0].split('/')[3] : '',
+  );
 
   const queryClient = useQueryClient();
 
@@ -49,17 +48,32 @@ export default function CommunityQuill({ queryParam }: CommunityQuillProps) {
       onSuccess: () => {
         alert('게시글을 수정하였습니다.');
         queryClient.refetchQueries([
-          QueryKeys.COMMUNITY_BOARD,
+          QueryKeys.INTRO_BOARD,
           `${boardData?.boardId}`,
         ]);
-        queryClient.refetchQueries([QueryKeys.COMMUNITY_BOARD]);
-        navigate(`/community/${queryParam}/${boardData?.boardId}`);
+        queryClient.refetchQueries([QueryKeys.INTRO_BOARD]);
+        navigate(`/intro_board/${boardData?.boardId}`);
       },
       onError: () => {
         alert('게시글 수정을 실패했습니다.');
       },
     },
   );
+
+  const thumbnailHandler = async (e: React.ChangeEvent<HTMLInputElement>) => {
+    if (e.currentTarget.files !== null) {
+      const file = e.currentTarget.files[0];
+      setThumbnailTitle(e.currentTarget.files[0].name);
+      try {
+        const res = await uploadFile(file);
+        const url = res || '';
+        setThumbnail(url);
+      } catch (error) {
+        const err = error as AxiosError;
+        return { ...err.response, success: false };
+      }
+    }
+  };
 
   // 이미지를 업로드 하기 위한 함수
   const modules = useQuillModules(QuillRef);
@@ -72,15 +86,15 @@ export default function CommunityQuill({ queryParam }: CommunityQuillProps) {
       title,
       code: contents,
       category,
-      imageUrls: [...getImageUrls(contents)],
-      prefixCategory,
+      imageUrls: [thumbnail, ...getImageUrls(contents)],
+      prefixCategory: 'INTRO',
       fixed: false,
     };
     const response = await PostBoardAPI(boardForm);
     if (response?.code === 'SUCCESS') {
-      alert('게시글이 작성되었습니다😄');
-      queryClient.refetchQueries([QueryKeys.COMMUNITY_BOARD]);
-      navigate(`/community/${queryParam}`);
+      alert('게시글이 작성되었습니다.');
+      queryClient.refetchQueries([QueryKeys.INTRO_BOARD]);
+      navigate('/introduce');
     }
   };
 
@@ -89,8 +103,8 @@ export default function CommunityQuill({ queryParam }: CommunityQuillProps) {
       title,
       code: contents,
       category,
-      imageUrls: [...getImageUrls(contents)],
-      prefixCategory,
+      imageUrls: [thumbnail, ...getImageUrls(contents)],
+      prefixCategory: 'INTRO',
       fixed: false,
     };
     mutate(boardForm);
@@ -107,11 +121,12 @@ export default function CommunityQuill({ queryParam }: CommunityQuillProps) {
 
   return (
     <div className={styles.container}>
-      <Title category={queryParam} boardData={boardData} />
+      <h1>{boardData ? '관리자 글 수정하기' : '관리자 글쓰기'}</h1>
       <div className={styles.sectionWrapper}>
         <section className={styles.labelSection}>
           <label>말머리</label>
           <label>제목</label>
+          <label>썸네일</label>
           <label>내용</label>
         </section>
         <section className={styles.contentSection}>
@@ -126,11 +141,8 @@ export default function CommunityQuill({ queryParam }: CommunityQuillProps) {
             <option disabled value="">
               말머리를 선택하세요.
             </option>
-            {categoryList.map((item) => (
-              <option key={item.name} value={item.name}>
-                {item.code}
-              </option>
-            ))}
+            <option value="TREND">트렌드</option>
+            <option value="REVIEW">후기</option>
           </select>
           <input
             className={styles.titleInput}
@@ -141,7 +153,23 @@ export default function CommunityQuill({ queryParam }: CommunityQuillProps) {
               setTitle(e.target.value)
             }
           />
-
+          <div className={styles.thumbnailWrapper}>
+            <input
+              type="text"
+              value={thumbnailTitle}
+              placeholder="사진을 첨부해주세요. "
+              readOnly
+            />
+            <input
+              ref={thumbnailRef}
+              style={{ display: 'none' }}
+              type="file"
+              onChange={thumbnailHandler}
+            />
+            <button type="button" onClick={() => thumbnailRef.current?.click()}>
+              업로드
+            </button>
+          </div>
           <ReactQuill
             className={styles.quill}
             ref={(element) => {
@@ -154,17 +182,15 @@ export default function CommunityQuill({ queryParam }: CommunityQuillProps) {
           />
         </section>
       </div>
-      <section>
-        {boardData ? (
-          <button type="button" onClick={onUpdate}>
-            수정하기
-          </button>
-        ) : (
-          <button type="button" onClick={onPost}>
-            등록하기
-          </button>
-        )}
-      </section>
+      {boardData ? (
+        <button type="button" onClick={onUpdate}>
+          수정하기
+        </button>
+      ) : (
+        <button type="button" onClick={onPost}>
+          등록하기
+        </button>
+      )}
     </div>
   );
 }
